fix expand_as_v2 test
